+++ src/components/appbar/constants/header.ts
@@ -1,0 +1,2 @@
+export const HOME_TRENDING_SENTINEL_ID = "home-trending-sentinel";
+export const APPBAR_HEIGHT = 73;
