+++ src/components/appbar/index.ts
@@ -1,0 +1,1 @@
+export { AppBar } from "@/components/appbar/AppBar";
