+++ src/components/appbar/types/appbar.ts
@@ -1,0 +1,9 @@
+import type { ViewType } from "@/routes/routes";
+
+export interface AppBarProps {
+  title: string;
+  currentView: ViewType;
+  onNavigate: (view: ViewType) => void;
+  pathname?: string;
+  breakpoint: "mobile" | "tablet" | "desktop";
+}
